FINALLY!
A first stab at a CW version based on EVE. Warning: This is very much work in progress and has not been exhaustively tested although the main functions with the exception of the MM seem to work alright. (You have been warned!) The moving map is not functional and still needs extensive work due to the many changes of the graphics framework.

The current version should (more or less) be equivalent to CW-EWE-1533. Some additional changes:
- Naming convention changed in-line with Java conventions (directories in lowercase, Classes in Uppercase)
- Also some tidying within files: Class member names changed to lowercase.
- Four subdirectories to cachewolf: exp (=exporters), imp (=importers including spider), navi (=moving map) and utils (=utilities).
- Some related files were combined to make editing easier, e.g. TablePanel now contains old myTableModel and myTableControl and myComparer, Parser now contains Tokenizer etc. The cachewolf directory should only contain screens and main classes - there are still too many files in this directory.
- References to FileBugfix were removed - need to test whether the file system bugs have been fixed in EVE. If necessary, EVE needs to be updated.
- All helper files (language files, attributes, images) are identical to CW-EWE.
- The "DelayedDisplay" to speed up loading on the PDA which was suggested by Mike Brereton in the EVE forum has not been implemented yet.
- RadarPanel background is buffered for faster redrawing when a different cache is selected in the listview
- some other changes which I have now forgotten :-(

Extensive testing now needed. I did not have time for this as pressure to release some code was mounting and I have been very busy at work over the last few months. The moving map is not functional. With hindsight it was a mistake to start the conversion to EVE off-line, as many changes were checked in after I had done a considerable amount of work which I did not want to undo.  :-(

Could somebody who is more conversant with SVN (Engywuck?) please move these files to an appropriate location.

salzkammergut